[chore/#19] 패키징 정리
