+++ app/src/main/java/sopt/org/millie/presentation/search/bookdetail/component/book/BookDataSection.kt
@@ -30,8 +30,8 @@ import androidx.compose.ui.unit.dp
 import sopt.org.millie.R
 import sopt.org.millie.core.designsystem.theme.MillieTheme
 import sopt.org.millie.core.util.noRippleClickable
-import sopt.org.millie.presentation.bookdetail.component.tag.BookDetailTag
 import sopt.org.millie.presentation.bookdetail.model.BookDataType
+import sopt.org.millie.presentation.search.bookdetail.component.tag.BookDetailTag
 
 @Composable
 fun BookDataSection(
