+++ app/src/main/java/sopt/org/millie/presentation/search/bookdetail/component/book/BookSimilarSection.kt
@@ -22,6 +22,7 @@ import kotlinx.collections.immutable.persistentListOf
 import sopt.org.millie.R
 import sopt.org.millie.core.designsystem.theme.MillieTheme
 import sopt.org.millie.presentation.bookdetail.model.BookSimilarModel
+import sopt.org.millie.presentation.search.bookdetail.component.book.BookDetailBook
 
 @Composable
 fun BookSimilarSection(
