--- app/src/main/java/sopt/org/millie/presentation/search/searchresult/DummyScreen3.kt
@@ -1,1 +1,0 @@
-
