+++ app/src/main/java/sopt/org/millie/presentation/search/searchresult/book/component/SearchTitleAuthor.kt
@@ -1,4 +1,4 @@
-package sopt.org.millie.presentation.search.searchresult.component.book
+package sopt.org.millie.presentation.search.searchresult.book.component
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
