+++ app/src/main/java/sopt/org/millie/presentation/search/searchresult/book/model/SearchBannerModel.kt
@@ -1,4 +1,4 @@
-package sopt.org.millie.presentation.search.model
+package sopt.org.millie.presentation.search.searchresult.book.model
 import androidx.compose.runtime.Immutable
 
 @Immutable
