+++ app/src/main/java/sopt/org/millie/presentation/search/searchresult/post/component/SearchPostTitle.kt
@@ -1,4 +1,4 @@
-package sopt.org.millie.presentation.search.post
+package sopt.org.millie.presentation.search.searchresult.post.component
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
@@ -6,7 +6,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import sopt.org.millie.core.designsystem.theme.MillieTheme
-import sopt.org.millie.presentation.search.component.SearchTitle
+import sopt.org.millie.presentation.search.searchresult.component.SearchTitle
 
 @Composable
 fun SearchPostTitle(
